:art: Dividing Classes into Modules. Aggregate Modules into Packages
