--- src/CrawlerEngine/Parser/Downloaders.py
@@ -1,21 +1,0 @@
-import requests
-from abc import ABC, abstractmethod
-
-class IDownloader(ABC):
-    @abstractmethod
-    def download():
-        pass
-
-class DownloaderRequests(IDownloader):
-    def __init__(self):
-        pass
-    
-    def download():
-        pass
-
-class DownloaderSelenium(IDownloader):
-    def __init__(self):
-        pass
-    
-    def download():
-        pass
