--- src/CrawlerEngine/Parser/Parsers.py
@@ -1,51 +1,0 @@
-import Downloaders  
-from abc import ABC, abstractmethod
-
-class IParser(ABC):
-    @abstractmethod
-    def load_content():
-        pass
-
-    @abstractmethod
-    def parse():
-        pass
-
-class ParserJson(IParser):
-    def __init__(self):
-        pass
-
-    def load_content():
-        pass
-
-    def parse():
-        pass
-
-class ParserYaml(IParser):
-    def __init__(self):
-        pass
-
-    def load_content():
-        pass
-
-    def parse():
-        pass
-
-class ParserAtom(IParser):
-    def __init__(self):
-        pass
-
-    def load_content():
-        pass
-
-    def parse():
-        pass
-
-class ParserXml(IParser):
-    def __init__(self):
-        pass
-
-    def load_content():
-        pass
-
-    def parse():
-        pass
