+++ src/CveCrawler/CrawlerEngine/Parsers/Downloaders/DownloaderRequests.py
@@ -1,0 +1,8 @@
+from .IDownloader import IDownloader
+
+class DownloaderRequests(IDownloader):
+    def __init__(self):
+        pass
+    
+    def download():
+        pass
