+++ src/CveCrawler/CrawlerEngine/Parsers/Downloaders/DownloaderSelenium.py
@@ -1,0 +1,8 @@
+from .IDownloader import IDownloader
+
+class DownloaderSelenium(IDownloader):
+    def __init__(self):
+        pass
+    
+    def download():
+        pass
