+++ src/CveCrawler/CrawlerEngine/Parsers/Downloaders/IDownloader.py
@@ -1,0 +1,6 @@
+from abc import ABC, abstractmethod
+
+class IDownloader(ABC):
+    @abstractmethod
+    def download():
+        pass
