+++ src/CveCrawler/CrawlerEngine/Parsers/Downloaders/__init__.py
@@ -1,0 +1,7 @@
+from .DownloaderRequests import DownloaderRequests
+from .DownloaderSelenium import DownloaderSelenium
+
+__all__ = (
+    DownloaderRequests,
+    DownloaderSelenium,
+)
