+++ src/CveCrawler/CrawlerEngine/Parsers/IParser.py
@@ -1,0 +1,10 @@
+from abc import ABC, abstractmethod
+
+class IParser(ABC):
+    @abstractmethod
+    def load_content():
+        pass
+
+    @abstractmethod
+    def parse():
+        pass
