+++ src/CveCrawler/CrawlerEngine/Parsers/ParserAtom.py
@@ -1,0 +1,11 @@
+from .IParser import IParser
+
+class ParserAtom(IParser):
+    def __init__(self):
+        pass
+
+    def load_content():
+        pass
+
+    def parse():
+        pass
