+++ src/CveCrawler/CrawlerEngine/Parsers/ParserXml.py
@@ -1,0 +1,11 @@
+from .IParser import IParser
+
+class ParserXml(IParser):
+    def __init__(self):
+        pass
+
+    def load_content():
+        pass
+
+    def parse():
+        pass
