+++ src/CveCrawler/CrawlerEngine/Parsers/__init__.py
@@ -1,0 +1,11 @@
+from .ParserAtom import ParserAtom
+from .ParserJson import ParserJson
+from .ParserYaml import ParserYaml
+from .ParserXml  import ParserXml
+
+__all__ = (
+    ParserAtom,
+    ParserJson,
+    ParserYaml,
+    ParserXml,
+)
